Move time proptest generators into their own crate
Closes #828

This allows us to use the timestamp and DateTime generators in any other
crate in our workspace. It should help advance #822 and #821.
