+++ pbt-gen/README.md
@@ -1,0 +1,11 @@
+# PBT Gen
+
+This internal crate collects [proptest](https://github.com/AltSysrq/proptest)
+generators used for testing, are of some general utility, and don't rely on any
+code internal to the various crates of this repository.
+
+To view the documentation (which includes examples) run
+
+```sh
+cargo doc --open -p pbt-gen
+```
