+++ tendermint/Cargo.toml
@@ -68,3 +68,4 @@ secp256k1 = ["k256", "ripemd160"]
 [dev-dependencies]
 
 proptest = "0.10.1"
+pbt-gen = { path = "../pbt-gen" }
