+++ tendermint/src/time.rs
@@ -144,141 +144,9 @@ pub trait ParseTimestamp {
 
 #[cfg(test)]
 mod tests {
-    use std::convert::TryInto;
-
     use super::*;
-
-    // TODO(shon) Extract arbitrary generators into own library
-
-    use chrono::{DateTime, NaiveDate, TimeZone, Timelike, Utc};
+    use pbt_gen;
     use proptest::{prelude::*, sample::select};
-
-    // Any higher, and we're at seconds
-    const MAX_NANO_SECS: u32 = 999_999_999u32;
-
-    // With values larger or smaller then these, chrono produces invalid rfc3339
-    // timestamps. See https://github.com/chronotope/chrono/issues/537
-    fn min_time() -> DateTime<Utc> {
-        Utc.timestamp(-9999999999, 0)
-    }
-
-    fn max_time() -> DateTime<Utc> {
-        Utc.timestamp(99999999999, 0)
-    }
-
-    fn num_days_in_month(year: i32, month: u32) -> u32 {
-        // Using chrono, we get the duration beteween this month and the next,
-        // then count the number of days in that duration.  See
-        // https://stackoverflow.com/a/58188385/1187277
-        let given_month = NaiveDate::from_ymd(year, month, 1);
-        let next_month = NaiveDate::from_ymd(
-            if month == 12 { year + 1 } else { year },
-            if month == 12 { 1 } else { month + 1 },
-            1,
-        );
-        next_month
-            .signed_duration_since(given_month)
-            .num_days()
-            .try_into()
-            .unwrap()
-    }
-
-    prop_compose! {
-        /// An abitrary `chrono::DateTime` that is between `min` and `max`
-        /// DateTimes.
-        fn arb_datetime_in_range(min: DateTime<Utc>, max: DateTime<Utc>)(
-            secs in min.timestamp()..max.timestamp()
-        )(
-            // min mano secods is only relevant if we happen to hit the minimum
-            // seconds on the nose.
-            nano in (if secs == min.timestamp() { min.nanosecond() } else { 0 })..MAX_NANO_SECS,
-            // Make secs in scope
-            secs in Just(secs),
-        ) -> DateTime<Utc> {
-            println!(">> Secs {:?}", secs);
-            Utc.timestamp(secs, nano)
-        }
-    }
-
-    prop_compose! {
-        /// An abitrary `chrono::DateTime`
-        fn arb_datetime()
-            (
-                d in arb_datetime_in_range(min_time(), max_time())
-            ) -> DateTime<Utc> {
-                d
-            }
-    }
-
-    prop_compose! {
-        fn arb_rfc339_time_offset()(
-            sign in "[+-]",
-            hour in 0..23u8,
-            min in 0..59u8,
-        ) -> String {
-            format!("{:}{:0>2}:{:0>2}", sign, hour, min)
-        }
-    }
-
-    fn arb_rfc3339_offset() -> impl Strategy<Value = String> {
-        prop_oneof![arb_rfc339_time_offset(), Just("Z".to_owned())]
-    }
-
-    prop_compose! {
-        fn arb_rfc3339_partial_time()(
-            hour in 0..23u8,
-            min in 0..59u8,
-            sec in 0..59u8,
-            secfrac in proptest::option::of(0..u64::MAX),
-        ) -> String {
-            let frac = match secfrac {
-                None => "".to_owned(),
-                Some(frac) => format!(".{:}", frac)
-            };
-            format!("{:0>2}:{:0>2}:{:0>2}{:}", hour, min, sec, frac)
-        }
-    }
-
-    prop_compose! {
-        fn arb_rfc3339_full_time()(
-            time in arb_rfc3339_partial_time(),
-            offset in arb_rfc3339_offset()
-        ) -> String {
-            format!("{:}{:}", time, offset)
-        }
-    }
-
-    prop_compose! {
-        fn arb_rfc3339_day_of_year_and_month(year: i32, month: u32)
-            (
-                d in 1..num_days_in_month(year, month)
-            ) -> u32 {
-            d
-        }
-    }
-
-    prop_compose! {
-        fn arb_rfc3339_full_date()(year in 0..9999i32, month in 1..12u32)
-            (
-                day in arb_rfc3339_day_of_year_and_month(year, month),
-                year in Just(year),
-                month in Just(month),
-            ) -> String {
-                format!("{:0>4}-{:0>2}-{:0>2}", year, month, day)
-            }
-    }
-
-    prop_compose! {
-        /// An aribtrary rfc3339 timestamp
-        ///
-        /// Follows https://tools.ietf.org/html/rfc3339#section-5.6
-        fn arb_rfc3339_timestamp()(
-            date in arb_rfc3339_full_date(),
-            time in arb_rfc3339_full_time()
-        ) -> String {
-            format!("{:}T{:}", date, time)
-        }
-    }
 
     // We want to make sure that these timestamps specifically get tested.
     fn particular_rfc3339_timestamps() -> impl Strategy<Value = String> {
@@ -309,13 +177,13 @@ mod tests {
 
     proptest! {
         #[test]
-        fn can_parse_rfc3339_timestamps(stamp in arb_rfc3339_timestamp()) {
+        fn can_parse_rfc3339_timestamps(stamp in pbt_gen::time::arb_rfc3339_timestamp()) {
             prop_assert!(stamp.parse::<Time>().is_ok())
         }
 
         #[test]
         fn serde_from_value_is_the_inverse_of_to_value_within_reasonable_time_range(
-            datetime in arb_datetime()
+            datetime in pbt_gen::time::arb_datetime()
         ) {
             // If `from_value` is the inverse of `to_value`, then it will always
             // map the JSON `encoded_time` to back to the inital `time`.
@@ -328,7 +196,7 @@ mod tests {
         #[test]
         fn serde_of_rfc3339_timestamps_is_safe(
             stamp in prop_oneof![
-                arb_rfc3339_timestamp(),
+                pbt_gen::time::arb_rfc3339_timestamp(),
                 particular_rfc3339_timestamps(),
             ]
         ) {
